BUG-58: Introduce new CVS exporter method and solve spring init. Update to spring 3.3.4

+++ src/lince-data-fx/src/main/java/com/lince/observer/legacy/FilaRegistro.java
@@ -147,7 +147,7 @@ public class FilaRegistro implements Comparable<FilaRegistro>, Serializable {
     }
 
     public int getRegisterFrameValue() {
-       return timeCalculations.convertMsToFPS(milis, this.fps);
+       return (int) timeCalculations.convertMsToFPS(milis, this.fps);
     }
 
     /**
